+++ digicampipe/scripts/baseline_shift.py
@@ -21,14 +21,15 @@ Options:
   --crosstalk=<CROSSTALK>     Calibration params to use in the fit
 """
 import os
-from docopt import docopt
-import numpy as np
+
 import matplotlib.pyplot as plt
-from tqdm import tqdm
+import numpy as np
+from docopt import docopt
 from histogram.histogram import Histogram1D
+from tqdm import tqdm
 
 from digicampipe.io.event_stream import calibration_event_stream
-from digicampipe.utils.docopt import convert_max_events_args, \
+from digicampipe.utils import convert_max_events_args, \
     convert_pixel_args, convert_dac_level
