+++ digicampipe/tests/test_dataquality.py
@@ -5,7 +5,7 @@ from pkg_resources import resource_filename
 from astropy.io import fits
 from digicampipe.scripts.data_quality import entry as data_quality
 from digicampipe.scripts.raw import compute as compute_raw
-from digicampipe.utils.docopt import convert_pixel_args
+from digicampipe.utils import convert_pixel_args
 
 
 example_file1_path = resource_filename(
