+++ digicampipe/tests/test_pipeline.py
@@ -1,10 +1,12 @@
-import tempfile
 import os
-from pkg_resources import resource_filename
+import tempfile
+
 from astropy.io import fits
+from pkg_resources import resource_filename
+
 from digicampipe.scripts.pipeline import main as main_pipeline
 from digicampipe.scripts.raw import compute as compute_raw
-from digicampipe.utils.docopt import convert_pixel_args
+from digicampipe.utils import convert_pixel_args
 
 example_file1_path = resource_filename(
     'digicampipe',
